begin profile

+++ app/src/main/java/gg/solrudev/vitaland/domain/model/ShiftRating.kt
@@ -3,6 +3,6 @@ package gg.solrudev.vitaland.domain.model
 data class ShiftRating(
 	val user: User,
 	val shift: Shift,
-	val rating: Double,
+	val rating: Int,
 	val text: String
 )
